requirements.txt
pyyaml==6.0.2
requests==2.32.3
